Replace setProotocolAndRelayerFeesUSDC.js with setProotocolFees.js, updating the fee management logic to utilize the FeeCollector contract. The new script sets platform and private fees for USDT and ONE addresses, enhancing fee configuration through environment variables.
